Extract start crawl mutation mock functionality into utilts.tsx to reduce duplication
