+++ ui/src/components/__tests__/helpers/utils.tsx
@@ -3,6 +3,7 @@ import { MockedProvider, MockedResponse } from "@apollo/client/testing";
 import { render } from "@testing-library/react";
 
 import { CRAWL_STATUS_UPDATE_SUBSCRIPTION } from "../../CrawlingSpinner";
+import { START_CRAWL_MUTATION } from "../../Search";
 import CrawlStatus from "../../../enums/CrawlStatus";
 
 function renderWithMockProvider(
@@ -30,4 +31,18 @@ function createStatusUpdateMock(mockStatus: CrawlStatus, expectedURL: string) {
     };
 }
 
-export { renderWithMockProvider, createStatusUpdateMock };
+function createStartCrawlMock(started: boolean, url: string) {
+    return {
+        request: {
+            query: START_CRAWL_MUTATION,
+            variables: { input: { url } },
+        },
+        result: jest.fn(() => ({
+            data: {
+                startCrawl: { started },
+            },
+        })),
+    };
+}
+
+export { renderWithMockProvider, createStatusUpdateMock, createStartCrawlMock };
